+++ limiter-tokenbucket/src/main/java/com/limiter/tokenbucket/config/TokenBucketConfig.java
@@ -38,7 +38,7 @@ public interface TokenBucketConfig {
      *
      * @return
      */
-    int getAddPeriod();
+    long getAddPeriod();
 
 
 }
